requirements.txt
# Pinned dependencies for reproducible builds
# Core ML
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
Pillow==10.4.0
opencv-python==4.10.0.84
ultralytics==8.1.0

# Grad-CAM
pytorch-grad-cam==1.5.0

# XGBoost
xgboost==2.0.3

# Configuration
PyYAML==6.0.1

# Desktop GUI
PySide6==6.7.2
